Move Annotation repository to Core
