+++ src/Wallabag/CoreBundle/Controller/AnnotationController.php
@@ -12,9 +12,9 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 use Wallabag\AnnotationBundle\Form\EditAnnotationType;
 use Wallabag\AnnotationBundle\Form\NewAnnotationType;
-use Wallabag\AnnotationBundle\Repository\AnnotationRepository;
 use Wallabag\CoreBundle\Entity\Annotation;
 use Wallabag\CoreBundle\Entity\Entry;
+use Wallabag\CoreBundle\Repository\AnnotationRepository;
 use Wallabag\UserBundle\Entity\User;
 
 class AnnotationController extends AbstractFOSRestController
